fix: Correctly recurse relative paths on file and href
Previously all file paths were relative to the CWD, even those that were on remote servers or NPM modules, or even in other folders. Now file paths are relative to the file that is doing the include.

This isn't considered a breaking change as the previous behavior wasn't an expected behavior and was indeed wrong. I really need to make automated tests.